Small changes and fix of fatal error when training cbs
Small changes made during familiarization with changes (cosmetic changes, correction of warnings)

Fix of fatal error when trying to train cbs.
Cause wasn't the usage of all cases but an error when loading the datasets, in case the loaded dataset did not contain a required feature. Error case is now checked and throws exception.