Preparations for FastCBS Variants
Changed the way of encoding for snn, modifed version will be used by cbs.
The encoding wont be initialized by the dataset itself but is now a method of the dataset that is called by the constructor of the fast variants
Also the encoding wont be stored in a file anymore, it just overwrites the loaded dataset.

Visualisation has been adapted but still needs to tested